Add Ui, Parser, DateTime, Storage and TaskList class
