+++ data/eren.txt
@@ -1,4 +1,5 @@
   TYPE  | COMPLETED | DETAILS | DATE
-Deadline|    YES    | hoho | 27 Jan 1999 03:00pm
-  Event |    NO     | bobo | 27 Jan 1999 07:00pm - 28 Jan 1999 09:00am
-  Todo  |    NO     | hasdhasdas | 27 Jan 1999 07:00pm - 28 Jan 1999 09:00am
+Deadline|    NO     | hoho | 27 Jan 1999 03:00pm
+  Event |    YES    | bobo | 27 Jan 1999 07:00pm - 28 Jan 1999 09:00am
+  Todo  |    NO     | asdasd | 27 Jan 1999 07:00pm - 28 Jan 1999 09:00am
+Deadline|    NO     | hi | 356 Dec 2022 10:00am
